Automated the project setup process.
The database password will be taken from now on from the `.env` file, along with an added `requirements.txt` file for project-relevant requirements and a `setup.py` file that will create a virtual environment, install requirements, build the Docker image and the container automatically, and run the backend script at the end while also checking for existing virtual environments or Docker containers. A `.gitignore` file was created to ignore created virtual environments automatically as well.